pyproject.toml
[project]
name = "medilink"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.3",
    "django",
    "psycopg[binary]>=3.3.2",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "sqlfluff>=3.5.0",
]
